DHFPROD-4519: Add ModelManager test
